extract and test generic form data extractor
